+++ R/io_parquet.R
@@ -1,21 +1,19 @@
 #' Scan a parquet file
 #'
+#' @inherit pl_scan_csv return
 #' @inheritParams pl_scan_csv
 #' @param source Path to a file. You can use globbing with `*` to scan/read multiple
 #' files in the same directory (see examples).
 #' @param cache Cache the result after reading.
 #' @param parallel This determines the direction of parallelism. `"auto"` will
-#' try to determine the optimal direction. Can be `"auto"`, `"none"`, `"columns"`,
-#' or `"rowgroups"`,
+#' try to determine the optimal direction. Can be `"auto"`, `"columns"`,
+#' `"row_groups"`, or `"none"`.
 #' @param rechunk In case of reading multiple files via a glob pattern, rechunk
 #' the final DataFrame into contiguous memory chunks.
 #' @param hive_partitioning Infer statistics and schema from hive partitioned URL
 #' and use them to prune reads.
 #' @param use_statistics Use statistics in the parquet file to determine if pages
 #' can be skipped from reading.
-#'
-#' @return LazyFrame
-#' @name scan_parquet
 #' @rdname IO_scan_parquet
 #' @examplesIf requireNamespace("arrow", quietly = TRUE) && arrow::arrow_with_dataset() && arrow::arrow_with_parquet()
 #' temp_dir = tempfile()
@@ -37,19 +35,19 @@ pl_scan_parquet = function(
     source,
     ...,
     n_rows = NULL,
-    cache = TRUE,
-    parallel = c(
-      "Auto", # default
-      "None",
-      "Columns",
-      "RowGroups"
-    ),
-    rechunk = TRUE,
     row_index_name = NULL,
     row_index_offset = 0L,
+    parallel = c(
+      "auto",
+      "columns",
+      "row_groups",
+      "none"
+    ),
+    hive_partitioning = TRUE,
+    rechunk = FALSE,
     low_memory = FALSE,
     use_statistics = TRUE,
-    hive_partitioning = TRUE) {
+    cache = TRUE) {
   new_from_parquet(
     path = source,
     n_rows = n_rows,
@@ -58,7 +56,6 @@ pl_scan_parquet = function(
     rechunk = rechunk,
     row_name = row_index_name,
     row_index = row_index_offset,
-    # storage_options = storage_options, # not supported yet
     low_memory = low_memory,
     use_statistics = use_statistics,
     hive_partitioning = hive_partitioning
@@ -68,26 +65,41 @@ pl_scan_parquet = function(
 
 #' Read a parquet file
 #' @rdname IO_read_parquet
-#' @inheritParams scan_parquet
-#' @return DataFrame
+#' @inherit pl_read_csv return
+#' @inheritParams pl_scan_parquet
+#' @examplesIf requireNamespace("arrow", quietly = TRUE) && arrow::arrow_with_dataset() && arrow::arrow_with_parquet()
+#' temp_dir = tempfile()
+#' # Write a hive-style partitioned parquet dataset
+#' arrow::write_dataset(
+#'   mtcars,
+#'   temp_dir,
+#'   partitioning = c("cyl", "gear"),
+#'   format = "parquet",
+#'   hive_style = TRUE
+#' )
+#' list.files(temp_dir, recursive = TRUE)
+#'
+#' # Read the dataset
+#' pl$read_parquet(
+#'   file.path(temp_dir, "**/*.parquet")
+#' )
 pl_read_parquet = function(
-    # remapped to pl$read_parquet, a hack to support roxygen2 @inheritsParams
     source,
     ...,
     n_rows = NULL,
-    cache = TRUE,
-    parallel = c(
-      "Auto", # default
-      "None",
-      "Columns",
-      "RowGroups"
-    ),
-    rechunk = TRUE,
     row_index_name = NULL,
     row_index_offset = 0L,
+    parallel = c(
+      "auto",
+      "columns",
+      "row_groups",
+      "none"
+    ),
+    hive_partitioning = TRUE,
+    rechunk = TRUE,
     low_memory = FALSE,
     use_statistics = TRUE,
-    hive_partitioning = TRUE) {
+    cache = TRUE) {
   .args = as.list(environment())
   result({
     do.call(pl$scan_parquet, .args)$collect()